[SMX4-815]upgrade to cxf 2.4.0
git-svn-id: https://svn.apache.org/repos/asf/servicemix/smx4/features/trunk@1097694 13f79535-47bb-0310-9956-ffa450edef68

+++ examples/cxf-ws-rm/src/main/java/org/apache/servicemix/samples/cxf_ws_rm/MessageLossSimulator.java
@@ -106,7 +106,7 @@ public class MessageLossSimulator extends AbstractPhaseInterceptor<Message> {
         @Override
         protected void onFirstWrite() throws IOException {
             if (LOG.isLoggable(Level.FINE)) {
-                BigInteger nr = RMContextUtils.retrieveRMProperties(outMessage, true)
+                Long nr = RMContextUtils.retrieveRMProperties(outMessage, true)
                     .getSequence().getMessageNumber();
                 LOG.fine("Losing message " + nr);
             }
